refactor[ia3]: moved rabbitmq connection to separate class
