+++ HealthcareSystem/Backend/Communication/RabbitMqConnection/IRabbitMqConnection.cs
@@ -1,0 +1,11 @@
+﻿using Backend.Service;
+using RabbitMQ.Client;
+
+namespace Backend.Communication.RabbitMqConnection
+{
+    public interface IRabbitMqConnection
+    {
+        RabbitMqConfiguration Configuration { get; }
+        IConnection Connection { get; } 
+    }
+}
