+++ HealthcareSystem/Backend/Communication/RabbitMqConnection/RabbitMqConfiguration.cs
@@ -8,5 +8,7 @@
         public string Password { get; set; }
         public int RetryCount { get; set; }
         public int RetryWait { get; set; }
+        public string ActionBenefitQueueName { get; set; }
+        public string TenderQueueName { get; set; }
     }
 }
