+++ HealthcareSystem/Backend/Service/Tendering/RabbitMqTenderingService/ITenderingBackgroundService.cs
@@ -1,0 +1,6 @@
+﻿namespace Backend.Service.Tendering.RabbitMqTenderingService
+{
+    public interface ITenderingBackgroundService
+    {
+    }
+}
