+++ HealthcareSystem/IntegrationAdapters/Controllers/PharmacyController.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using Backend.Model.Pharmacies;
 using Backend.Service;
 using Backend.Service.Pharmacies;
-using IntegrationAdapters.Dtos;
 using Microsoft.AspNetCore.Mvc;
 
 namespace IntegrationAdapters.Controllers
@@ -11,12 +10,12 @@ namespace IntegrationAdapters.Controllers
     public class PharmacyController : Microsoft.AspNetCore.Mvc.Controller
     {
         private readonly IPharmacyService _pharmacyService;
-        private readonly IActionBenefitMessageingService _actionBenefitMessageingService;
+        private readonly IActionBenefitSubscriptionService _actionBenefitMessageingService;
 
-        public PharmacyController(IPharmacyService iPharmacyService, RabbitMqActionBenefitMessageingService actionBenefitMessageingService)
+        public PharmacyController(IPharmacyService iPharmacyService, IActionBenefitSubscriptionService actionBenefitSubscriptionService)
         {
             _pharmacyService = iPharmacyService;
-            _actionBenefitMessageingService = actionBenefitMessageingService;
+            _actionBenefitMessageingService = actionBenefitSubscriptionService;
         }
 
         public IActionResult ApiRegister()
