+++ HealthcareSystem/IntegrationAdapters/Startup.cs
@@ -31,6 +31,7 @@ using Backend.Repository.TenderRepository.MySqlTenderRepository;
 using Backend.Service.DrugAndTherapy;
 using Backend.Service.Tendering;
 using Service.DrugAndTherapy;
+using Backend.Communication.RabbitMqConnection;
 
 namespace IntegrationAdapters
 {
@@ -96,9 +97,10 @@ namespace IntegrationAdapters
             services.Configure<RabbitMqConfiguration>(Configuration.GetSection("RabbitMq"));
             services.Configure<RabbitMqConfiguration>(GetRabbitConfig);
             services.Configure<SftpConfig>(Configuration.GetSection("SftpConfig"));
-            services.AddSingleton<RabbitMqActionBenefitMessageingService>();
-            services.AddSingleton<IHostedService, RabbitMqActionBenefitMessageingService>(ServiceProvider => ServiceProvider.GetService<RabbitMqActionBenefitMessageingService>());
+            services.AddSingleton<IRabbitMqConnection, RabbitMqConnection>();
+            services.AddHostedService<RabbitMqActionBenefitMessageingService>();
 
+            services.AddScoped<IActionBenefitSubscriptionService, ActionBenefitSubscriptionService>();
             services.AddScoped<IPharmacyRepo, MySqlPharmacyRepo>();
             services.AddScoped<IPharmacyService, PharmacyService>();
             services.AddScoped<IActionBenefitRepository, MySqlActionBenefitRepository>();
@@ -129,6 +131,8 @@ namespace IntegrationAdapters
             conf.Password = Configuration.GetValue<string>("RABBITMQ_PASSWORD") ?? "guest";
             conf.RetryCount = Configuration.GetValue<int>("RABBITMQ_RETRY");
             conf.RetryWait = Configuration.GetValue<int>("RABBITMQ_RETRY_WAIT");
+            conf.ActionBenefitQueueName = Configuration.GetValue<string>("ACTIONBENEFIT_QUEUE") ?? "ab-bolnica-1";
+            conf.TenderQueueName = Configuration.GetValue<string>("TENDER_QUEUE") ?? "t-bolnica-1";
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
